Move Plate into components
